% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\alias{db_connect_local}
\title{Database Connection}
\usage{
db_connect(db_config = NULL, ...)

db_connect_local(db_config = NULL, ...)
}
\arguments{
\item{db_config}{A list containing the database connection configuration.}

\item{...}{Additional arguments to pass to \code{DBI::dbConnect}.}
}
\value{
A database connection object.
}
\description{
Connect to a database using the provided configuration.
}
\section{Functions}{
\itemize{
\item \code{db_connect_local()}: Connect to an local, in-memory database

}}
