% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate-package.R
\docType{package}
\name{migrate-package}
\alias{migrate}
\alias{migrate-package}
\title{migrate: Database Migrations in R}
\description{
Provides a database migration framework for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/migrate}
  \item \url{http://docs.jimbrig.com/migrate/}
  \item Report bugs at \url{https://github.com/jimbrig/migrate/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
