% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\docType{package}
\name{migrate}
\alias{migrate}
\alias{migrate-package}
\title{migrate: A package to estimate the migration rate between two areas from
tagging and population data}
\description{
The migrate package implements a modification of Hilborn (1990) conditioned on
catch instead of effort.
}
\section{Vignettes}{

To learn more about migrate, start with the vignettes:
\code{browseVignettes(package = "migrate")}
}

