% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_sim.R
\name{move_sim}
\alias{move_sim}
\title{Estimate movement between two areas from tag recaptures}
\usage{
move_sim(obs1, para, rfit, lltype)
}
\arguments{
\item{obs1}{this is the observations}

\item{para}{this is the parameters}

\item{rfit}{this is the rfit}

\item{lltype}{the fitting method (least squares or maximum likelihood)}
}
\description{
Estimate movement between two areas from tag recaptures, catches and
population sizes in each area.
}
